Update CameraLauncher.java
In android 4.4 if you tried to select an existing photo and didn't choose the "Gallery" application. Let's say you chose the "Photos" application then didn't select anything and hit the back button, the application that uses the plugin crashes. So we're adding the try/catch block to allow it to fail gracefully.